backend/python/chinese-english-readers/requirements.txt
sounddevice
vosk
gtts
playsound
edge-tts
pydub
fpdf2
pysmbclient
smbprotocol
